# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XogsideIQ - Crypto Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /narratives/{slug}:
    get:
      operationId: getNarrative
      tags: [narratives]
      summary: Get narrative detail with top tokens
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Narrative detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NarrativeDetail"
        "404":
          description: Narrative not found
  /signals:
    get:
      operationId: listSignals
      tags: [signals]
      summary: List trading signals
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, hit_target, stopped, expired]
          description: Filter by status
      responses:
        "200":
          description: Signals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Signal"
    post:
      operationId: createSignal
      tags: [signals]
      summary: Create a new trading signal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignalBody"
      responses:
        "201":
          description: Signal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
  /signals/{id}:
    patch:
      operationId: updateSignal
      tags: [signals]
      summary: Update a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSignalBody"
      responses:
        "200":
          description: Updated signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Not found
    delete:
      operationId: deleteSignal
      tags: [signals]
      summary: Delete a signal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /portfolio/positions:
    get:
      operationId: listPositions
      tags: [portfolio]
      summary: List portfolio positions
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
    post:
      operationId: createPosition
      tags: [portfolio]
      summary: Add a position to portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePositionBody"
      responses:
        "201":
          description: Position created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
  /portfolio/positions/{id}:
    patch:
      operationId: updatePosition
      tags: [portfolio]
      summary: Update a portfolio position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePositionBody"
      responses:
        "200":
          description: Updated position
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Position"
        "404":
          description: Not found
    delete:
      operationId: deletePosition
      tags: [portfolio]
      summary: Delete a portfolio position
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Token:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
        name:
          type: string
        logoUrl:
          type: ["string", "null"]
        chain:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        priceChange24h:
          type: ["number", "null"]
        marketCap:
          type: ["number", "null"]
        volume24h:
          type: ["number", "null"]
        overallScore:
          type: ["integer", "null"]
        finalGrade:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, symbol, name, chain, createdAt]
...
    NarrativeDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        iconKey:
          type: string
        momentumScore:
          type: integer
        perf24h:
          type: number
        perf7d:
          type: number
        aiCommentary:
          type: ["string", "null"]
        topTokens:
          type: array
          items:
            $ref: "#/components/schemas/Token"
        createdAt:
          type: string
      required: [id, name, slug, iconKey, momentumScore, perf24h, perf7d, topTokens, createdAt]
    Signal:
      type: object
      properties:
        id:
          type: integer
        tokenSymbol:
          type: string
        tokenName:
          type: string
        action:
          type: string
          enum: [BUY, SELL, WATCH]
        entryPrice:
          type: number
        targetPrice:
          type: number
        stopLossPrice:
          type: ["number", "null"]
        confidence:
          type: integer
        timeframe:
          type: string
          enum: [scalp, swing, position]
        status:
          type: string
          enum: [active, hit_target, stopped, expired]
        progressPercent:
          type: number
        createdAt:
          type: string
        expiresAt:
          type: ["string", "null"]
      required: [id, tokenSymbol, tokenName, action, entryPrice, targetPrice, confidence, timeframe, status, progressPercent, createdAt]
    CreateSignalBody:
      type: object
      properties:
        tokenSymbol:
          type: string
        tokenName:
          type: string
        action:
          type: string
          enum: [BUY, SELL, WATCH]
        entryPrice:
          type: number
        targetPrice:
          type: number
        stopLossPrice:
          type: ["number", "null"]
        confidence:
          type: integer
        timeframe:
          type: string
          enum: [scalp, swing, position]
        expiresAt:
          type: ["string", "null"]
      required: [tokenSymbol, tokenName, action, entryPrice, targetPrice, confidence, timeframe]
    UpdateSignalBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, hit_target, stopped, expired]
        targetPrice:
          type: number
        stopLossPrice:
          type: ["number", "null"]
        confidence:
          type: integer
    Position:
      type: object
      properties:
        id:
          type: integer
        tokenSymbol:
          type: string
        tokenName:
          type: string
        logoUrl:
          type: ["string", "null"]
        amount:
          type: number
        avgBuyPrice:
          type: number
        currentPrice:
          type: number
        valueUsd:
          type: number
        investedUsd:
          type: number
        pnlUsd:
          type: number
        pnlPercent:
          type: number
        targetPrice:
          type: ["number", "null"]
        targetProgressPercent:
          type: ["number", "null"]
        narrativeSlug:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, tokenSymbol, tokenName, amount, avgBuyPrice, currentPrice, valueUsd, investedUsd, pnlUsd, pnlPercent, createdAt]
    CreatePositionBody:
      type: object
      properties:
        tokenSymbol:
          type: string
        tokenName:
          type: string
        amount:
          type: number
        avgBuyPrice:
          type: number
        targetPrice:
          type: ["number", "null"]
        narrativeSlug:
          type: ["string", "null"]
      required: [tokenSymbol, tokenName, amount, avgBuyPrice]
    UpdatePositionBody:
      type: object
      properties:
        amount:
          type: number
        avgBuyPrice:
          type: number
        targetPrice:
          type: ["number", "null"]
...